test: consolidate core dispatch and async logic tests (fix lints)
