+++ tests/vibe3/commands/test_async_defaults.py
@@ -4,6 +4,7 @@ import re
 
 from typer.testing import CliRunner
 
+from vibe3.agents.runner import CodeagentExecutionService
 from vibe3.cli import app as cli_app
 
 runner = CliRunner(env={"NO_COLOR": "1"})
@@ -41,3 +42,22 @@ class TestAsyncDefaults:
         assert "sync" in output
         # Typer shows --no-async flag (inverted boolean)
         assert "--no-async" in output
+
+    def test_build_self_invocation_appends_no_async_for_tmux_child(self) -> None:
+        """Verify that self-invocation logic correctly standardizes on --no-async."""
+        cmd = CodeagentExecutionService.build_self_invocation(
+            ["run", "--plan", "/tmp/demo.md"]
+        )
+
+        assert "--no-async" in cmd
+        assert "--sync" not in cmd
+        assert "--async" not in cmd
+
+    def test_build_self_invocation_drops_legacy_async_and_standardizes(self) -> None:
+        """Verify that legacy --async flag is dropped and standardized to --no-async."""
+        cmd = CodeagentExecutionService.build_self_invocation(
+            ["review", "base", "--async"]
+        )
+
+        assert "--async" not in cmd
+        assert "--no-async" in cmd
